vuicode-app/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.1
pyyaml==6.0.1
starlette<0.37.0,>=0.36.3
openai==1.12.0
python-dotenv==1.0.1
markdownify==0.11.6
moviepy==1.0.3
elevenlabs==0.2.27
pytest==8.2.0
requests==2.32.3
httpx==0.27.0